fix: 공통된 getQueryClient 로직을 사용하여 쿼리 클라이언트 생성
- 서버와 브라우저 환경에서 쿼리 클라이언트를 통합된 로직으로 생성하도록 변경
